Move debug config handling to separate file
